Refactor update
Change `out` to `echo` since the former isn't installed on Fedora and exits on error when called.

Add a check to see if the script is running as root and exit if not.

Add relative references to sourced scripts as otherwise they fail to call and update exits on error.

Source new `update-snap` script.

Commented out `npm` for the time being.